chore: share the identifier rows with the mobile detail table
The helpers lived in `body.rs`, which is gated off the mobile build, so the
mobile metadata table couldn't reach them and still printed the raw scheme.
Moved them to an ungated `book_detail/identifiers.rs` and pointed both tables
at it; docs/architecture.md picks up the new module and the date trio.

Refs: #2251
Refs: #2243
